Adds listItemText as a styleable element
Currently, you can only style a `list`, a `listItem` or a `listItem`'s `listItemNumber` or `listItemBullet`, depending on whether it is ordered or not. Because a `listItem` is a `<View>` element, the only way to style the text inside it is to add styles to `text`. This overrides every other `text` element.